using layout for documentation to move out most of the code, but the md part
