+++ frontend/src/components/layout-types.ts
@@ -1,0 +1,7 @@
+import type { ReactElement, ReactNode } from 'react'
+import type { NextPage } from 'next'
+
+
+export type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
+    getLayout?: (page: ReactElement) => ReactNode
+}
